--- src/main/java/dev/entity/Admin.java
@@ -1,9 +1,0 @@
-package dev.entity;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "admin")
-public class Admin extends BaseEntity{
-
-}
