--- src/main/java/dev/entity/BaseEntity.java
@@ -1,18 +1,0 @@
-package dev.entity;
-
-import javax.persistence.*;
-
-@MappedSuperclass
-public class BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-}
