--- src/main/java/dev/entity/Favorite.java
@@ -1,32 +1,0 @@
-package dev.entity;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "favorite")
-public class Favorite extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id_city_fav", nullable = false)
-    private Long id1;
-
-    @Id
-    @Column(name = "id_user_fav", nullable = false)
-    private Long id2;
-
-    public Long getId2() {
-        return id2;
-    }
-
-    public void setId2(Long id2) {
-        this.id2 = id2;
-    }
-
-    public Long getId1() {
-        return id1;
-    }
-
-    public void setId1(Long id1) {
-        this.id1 = id1;
-    }
-}
