--- src/main/java/dev/entity/User.java
@@ -1,75 +1,0 @@
-package dev.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "user")
-public class User extends BaseEntity {
-    @Column(name = "username", nullable = false, length = 25)
-    private String username;
-
-    @Column(name = "firstname", nullable = false, length = 50)
-    private String firstname;
-
-    @Column(name = "lastname", nullable = false, length = 50)
-    private String lastname;
-
-    @Column(name = "mail", nullable = false, length = 50)
-    private String mail;
-
-    @Column(name = "password", nullable = false, length = 50)
-    private String password;
-
-    @Column(name = "active_account", nullable = false)
-    private Integer activeAccount;
-
-    public Integer getActiveAccount() {
-        return activeAccount;
-    }
-
-    public void setActiveAccount(Integer activeAccount) {
-        this.activeAccount = activeAccount;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getMail() {
-        return mail;
-    }
-
-    public void setMail(String mail) {
-        this.mail = mail;
-    }
-
-    public String getLastname() {
-        return lastname;
-    }
-
-    public void setLastname(String lastname) {
-        this.lastname = lastname;
-    }
-
-    public String getFirstname() {
-        return firstname;
-    }
-
-    public void setFirstname(String firstname) {
-        this.firstname = firstname;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-}
